initial version for exercise 23
It came down to using simple tree traversion to solve this, which might or might not be a good idea and will not scale properly if more questions are added and/or multiple options per questions are given. However, I determined that using an inference engine or FSM might be overkill here.